add separate folder for pojos
