+++ src/test/java/com/google/graphgeckos/dashboard/storage/DatastoreRepositoryTests.java
@@ -20,6 +20,8 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.concurrent.TimeoutException;
 import java.util.List;
+
+import com.google.graphgeckos.dashboard.datatypes.*;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
